BLE: remove unused test applications and update BLE HAL unit test application.
