+++ user/tests/app/ble/ibeacon/application.cpp
