+++ user/tests/app/ble/scanner/application.cpp
