+++ user/tests/app/ble/uart_central/application.cpp
