+++ user/tests/app/ble/uart_peripheral/application.cpp
