+++ user/tests/hal/ble/multi_roles/application.cpp
