+++ user/tests/hal/ble/multi_roles/multi_roles.cpp
@@ -24,9 +24,9 @@
 
 Serial1LogHandler log(115200, LOG_LEVEL_ALL);
 
-hal_ble_char_handles_t characteristic1; // Read and Write
-hal_ble_char_handles_t characteristic2; // Notify
-hal_ble_char_handles_t characteristic3; // Write without response
+hal_ble_char_handles_t localCharacteristic1; // Read and Write
+hal_ble_char_handles_t localCharacteristic2; // Notify
+hal_ble_char_handles_t localCharacteristic3; // Write without response
 
 const char* addrType[4] = {
     "Public",
@@ -39,9 +39,36 @@ uint8_t svc1UUID[] = {0x01,0x02,0x03,0x04,0x05,0x06,0x07,0x08,0x08,0x09,0x0a,0x0
 uint8_t char1UUID[] = {0x01,0x02,0x03,0x04,0x05,0x06,0x07,0x08,0x08,0x09,0x0a,0x0b,0x01,0x00,0x0e,0x0f};
 uint8_t svc2UUID[] = {0x01,0x02,0x03,0x04,0x05,0x06,0x07,0x08,0x08,0x09,0x0a,0x0b,0x00,0x00,0x0e,0x10};
 uint8_t char2UUID[] = {0x01,0x02,0x03,0x04,0x05,0x06,0x07,0x08,0x08,0x09,0x0a,0x0b,0x01,0x00,0x0e,0x10};
+uint16_t svc3UUID  = 0x1234;
+uint16_t char3UUID = 0x5678;
 
+uint16_t connHandle = BLE_INVALID_CONN_HANDLE;
+hal_ble_addr_t auxPeripheralAddr;
+bool auxPeripheralFound = false;
 
-static void ble_on_connected(hal_ble_gap_on_connected_evt_t* event) {
+hal_ble_svc_t service1, service2, service3;
+hal_ble_char_t char1, char2, char3;
+bool svc1Discovered = false, svc2Discovered = false, svc3Discovered= false;
+bool char1Discovered = false, char2Discovered = false, char3Discovered = false;
+bool notifiedData = false;
+
+static void ble_on_scan_result(const hal_ble_gap_on_scan_result_evt_t* event, void* context) {
+    uint8_t  devName[20];
+    uint16_t devNameLen = sizeof(devName);
+    if (event->data[4] == BLE_SIG_AD_TYPE_COMPLETE_LOCAL_NAME) {
+        devNameLen = event->data[3] - 1;
+        memcpy(devName, &event->data[5], devNameLen);
+        devName[devNameLen] = '\0';
+        LOG(TRACE, "Found device: %s", devName);
+        if (!strcmp((const char*)devName, "Xenon BLE Sample")) {
+            ble_gap_stop_scan();
+            auxPeripheralAddr = event->peer_addr;
+            auxPeripheralFound = true;
+        }
+    }
+}
+
+static void ble_on_connected(const hal_ble_gap_on_connected_evt_t* event) {
     LOG(TRACE, "BLE connected, connection handle: 0x%04X.", event->conn_handle);
     LOG(TRACE, "Local device role: %d.", event->role);
     if (event->peer_addr.addr_type <= 3) {
@@ -53,30 +80,80 @@ static void ble_on_connected(hal_ble_gap_on_connected_evt_t* event) {
     LOG(TRACE, "Peer address: %02X:%02X:%02X:%02X:%02X:%02X.", event->peer_addr.addr[0], event->peer_addr.addr[1],
                 event->peer_addr.addr[2], event->peer_addr.addr[3], event->peer_addr.addr[4], event->peer_addr.addr[5]);
     LOG(TRACE, "Interval: %.2fms, Latency: %d, Timeout: %dms", event->conn_interval*1.25, event->slave_latency, event->conn_sup_timeout*10);
+    connHandle = event->conn_handle;
 }
 
-static void ble_on_disconnected(hal_ble_gap_on_disconnected_evt_t* event) {
+static void ble_on_disconnected(const hal_ble_gap_on_disconnected_evt_t* event) {
     LOG(TRACE, "BLE disconnected, connection handle: 0x%04X.", event->conn_handle);
-    if (event->peer_addr.addr_type <= 3) {
-        LOG(TRACE, "Peer address type: %s", addrType[event->peer_addr.addr_type]);
-    }
-    else {
-        LOG(TRACE, "Peer address type: Anonymous");
-    }
-    LOG(TRACE, "Peer address: %02X:%02X:%02X:%02X:%02X:%02X.", event->peer_addr.addr[0], event->peer_addr.addr[1],
-                event->peer_addr.addr[2], event->peer_addr.addr[3], event->peer_addr.addr[4], event->peer_addr.addr[5]);
 }
 
-static void ble_on_data_received(hal_ble_gatt_on_data_evt_t* event) {
+static void ble_on_services_discovered(const hal_ble_gattc_on_svc_disc_evt_t* event, void* context) {
+    hal_ble_svc_t* service;
+    for (uint8_t i = 0; i < event->count; i++) {
+        service = NULL;
+        if (event->services[i].uuid.type == BLE_UUID_TYPE_16BIT) {
+            if (event->services[i].uuid.uuid16 == svc3UUID) {
+                service = &service3;
+                svc3Discovered = true;
+                LOG(TRACE, "BLE Service3 found.");
+            }
+        }
+        else if (event->services[i].uuid.type == BLE_UUID_TYPE_128BIT) {
+            if (!memcmp(svc1UUID, event->services[i].uuid.uuid128, BLE_SIG_UUID_128BIT_LEN)) {
+                service = &service1;
+                svc1Discovered = true;
+                LOG(TRACE, "BLE Service1 found.");
+            }
+            else if (!memcmp(svc2UUID, event->services[i].uuid.uuid128, BLE_SIG_UUID_128BIT_LEN)) {
+                service = &service2;
+                svc2Discovered = true;
+                LOG(TRACE, "BLE Service2 found.");
+            }
+        }
+        if (service != NULL) {
+            memcpy(service, &event->services[i], sizeof(hal_ble_svc_t));
+        }
+    }
+}
+
+static void ble_on_characteristics_discovered(const hal_ble_gattc_on_char_disc_evt_t* event, void* context) {
+    hal_ble_char_t* characteristic = NULL;
+    for (uint8_t i = 0; i < event->count; i++) {
+        if (event->characteristics[i].uuid.type == BLE_UUID_TYPE_16BIT) {
+            if (event->characteristics[i].uuid.uuid16 == char3UUID) {
+                characteristic = &char3;
+                char3Discovered = true;
+                LOG(TRACE, "BLE Characteristic3 found.");
+            }
+        }
+        else if (event->characteristics[i].uuid.type == BLE_UUID_TYPE_128BIT) {
+            if (!memcmp(char1UUID, event->characteristics[i].uuid.uuid128, BLE_SIG_UUID_128BIT_LEN)) {
+                characteristic = &char1;
+                char1Discovered = true;
+                LOG(TRACE, "BLE Characteristic1 found.");
+            }
+            else if (!memcmp(char2UUID, event->characteristics[i].uuid.uuid128, BLE_SIG_UUID_128BIT_LEN)) {
+                characteristic = &char2;
+                char2Discovered = true;
+                LOG(TRACE, "BLE Characteristic2 found.");
+            }
+        }
+        if (characteristic != NULL) {
+            memcpy(characteristic, &event->characteristics[i], sizeof(hal_ble_char_t));
+        }
+    }
+}
+
+static void ble_on_data_received(const hal_ble_gatt_on_data_evt_t* event) {
     LOG(TRACE, "BLE data received, connection handle: 0x%04X.", event->conn_handle);
 
-    if (event->attr_handle == characteristic1.value_handle) {
+    if (event->attr_handle == localCharacteristic1.value_handle) {
         LOG(TRACE, "Write BLE characteristic 1 value:");
     }
-    else if (event->attr_handle == characteristic2.cccd_handle) {
+    else if (event->attr_handle == localCharacteristic2.cccd_handle) {
         LOG(TRACE, "Configure BLE characteristic 2 CCCD:");
     }
-    else if (event->attr_handle == characteristic3.value_handle) {
+    else if (event->attr_handle == localCharacteristic3.value_handle) {
         LOG(TRACE, "Write BLE characteristic 3 value:");
     }
     else {
@@ -89,7 +166,14 @@ static void ble_on_data_received(hal_ble_gatt_on_data_evt_t* event) {
     Serial1.print("\r\n");
 }
 
-static void ble_on_events(hal_ble_evts_t *event, void* context) {
+static void ble_on_data_notified(const hal_ble_gatt_on_data_evt_t* event) {
+    if (event->attr_handle == char2.value_handle) {
+        LOG(TRACE, "Notified BLE characteristic 2 value:");
+        notifiedData = true;
+    }
+}
+
+static void ble_on_events(const hal_ble_evts_t *event, void* context) {
     if (event->type == BLE_EVT_CONNECTED) {
         ble_on_connected(&event->params.connected);
     }
@@ -99,18 +183,19 @@ static void ble_on_events(hal_ble_evts_t *event, void* context) {
     else if (event->type == BLE_EVT_DATA_WRITTEN) {
         ble_on_data_received(&event->params.data_rec);
     }
+    else if (event->type == BLE_EVT_DATA_NOTIFIED) {
+        ble_on_data_notified(&event->params.data_rec);
+    }
 }
 
 test(01_BleStackReinitializationShouldFail) {
     int ret;
 
     // First time to initialize BLE stack
-    ble_stack_init(NULL);
+    ret = ble_stack_init(nullptr);
+    assertEqual(ret, 0);
 
-    ret = ble_stack_init(NULL);
-    assertNotEqual(ret, 0);
-
-    ble_set_callback_on_events(ble_on_events, NULL);
+    ble_set_callback_on_events(ble_on_events, nullptr);
 }
 
 test(02_BleSetDeviceAddressShouldBePublicOrRandomStatic) {
@@ -197,23 +282,23 @@ test(05_BleSetPpcpShouldBeWithinValidRange) {
     setConnParams.max_conn_interval = BLE_SIG_CP_MAX_CONN_INTERVAL_NONE;
     setConnParams.slave_latency     = 1;
     setConnParams.conn_sup_timeout  = 400;
-    ret = ble_gap_set_ppcp(&setConnParams, NULL);
+    ret = ble_gap_set_ppcp(&setConnParams, nullptr);
     assertNotEqual(ret, 0);
 
     setConnParams.min_conn_interval = BLE_SIG_CP_MIN_CONN_INTERVAL_NONE;
     setConnParams.max_conn_interval = BLE_SIG_CP_MAX_CONN_INTERVAL_MAX + 1;
-    ret = ble_gap_set_ppcp(&setConnParams, NULL);
+    ret = ble_gap_set_ppcp(&setConnParams, nullptr);
     assertNotEqual(ret, 0);
 
     setConnParams.min_conn_interval = BLE_SIG_CP_MIN_CONN_INTERVAL_NONE;
     setConnParams.max_conn_interval = BLE_SIG_CP_MAX_CONN_INTERVAL_NONE;
     setConnParams.slave_latency     = BLE_SIG_CP_SLAVE_LATENCY_MAX + 1;
-    ret = ble_gap_set_ppcp(&setConnParams, NULL);
+    ret = ble_gap_set_ppcp(&setConnParams, nullptr);
     assertNotEqual(ret, 0);
 
     setConnParams.slave_latency     = 1;
     setConnParams.conn_sup_timeout  = BLE_SIG_CP_CONN_SUP_TIMEOUT_MAX + 1;
-    ret = ble_gap_set_ppcp(&setConnParams, NULL);
+    ret = ble_gap_set_ppcp(&setConnParams, nullptr);
     assertNotEqual(ret, 0);
 
     // Set valid PPCP
@@ -221,11 +306,11 @@ test(05_BleSetPpcpShouldBeWithinValidRange) {
     setConnParams.max_conn_interval = 400;
     setConnParams.slave_latency     = 1;
     setConnParams.conn_sup_timeout  = 400;
-    ret = ble_gap_set_ppcp(&setConnParams, NULL);
+    ret = ble_gap_set_ppcp(&setConnParams, nullptr);
     assertEqual(ret, 0);
 
     // Get PPCP
-    ret = ble_gap_get_ppcp(&getConnParams, NULL);
+    ret = ble_gap_get_ppcp(&getConnParams, nullptr);
     assertEqual(ret, 0);
     assertEqual(getConnParams.min_conn_interval, 100);
     assertEqual(getConnParams.max_conn_interval, 400);
@@ -241,47 +326,47 @@ test(06_BleSetTxPowerShouldBeRounded) {
     // Valid TX power for nRF52840: -20, -16, -12, -8, -4, 0, 4, 8
     ret = ble_gap_set_tx_power(-30);
     assertEqual(ret, 0);
-    ret = ble_gap_get_tx_power(&txPower);
+    txPower = ble_gap_get_tx_power();
     assertEqual(txPower, -20);
 
     ret = ble_gap_set_tx_power(-18);
     assertEqual(ret, 0);
-    ret = ble_gap_get_tx_power(&txPower);
+    txPower = ble_gap_get_tx_power();
     assertEqual(txPower, -16);
 
     ret = ble_gap_set_tx_power(-14);
     assertEqual(ret, 0);
-    ret = ble_gap_get_tx_power(&txPower);
+    txPower = ble_gap_get_tx_power();
     assertEqual(txPower, -12);
 
     ret = ble_gap_set_tx_power(-10);
     assertEqual(ret, 0);
-    ret = ble_gap_get_tx_power(&txPower);
+    txPower = ble_gap_get_tx_power();
     assertEqual(txPower, -8);
 
     ret = ble_gap_set_tx_power(-6);
     assertEqual(ret, 0);
-    ret = ble_gap_get_tx_power(&txPower);
+    txPower = ble_gap_get_tx_power();
     assertEqual(txPower, -4);
 
     ret = ble_gap_set_tx_power(-2);
     assertEqual(ret, 0);
-    ret = ble_gap_get_tx_power(&txPower);
+    txPower = ble_gap_get_tx_power();
     assertEqual(txPower, 0);
 
     ret = ble_gap_set_tx_power(2);
     assertEqual(ret, 0);
-    ret = ble_gap_get_tx_power(&txPower);
+    txPower = ble_gap_get_tx_power();
     assertEqual(txPower, 4);
 
     ret = ble_gap_set_tx_power(6);
     assertEqual(ret, 0);
-    ret = ble_gap_get_tx_power(&txPower);
+    txPower = ble_gap_get_tx_power();
     assertEqual(txPower, 8);
 
     ret = ble_gap_set_tx_power(10);
     assertEqual(ret, 0);
-    ret = ble_gap_get_tx_power(&txPower);
+    txPower = ble_gap_get_tx_power();
     assertEqual(txPower, 8);
 #endif
 }
@@ -294,10 +379,10 @@ test(07_BleSetAdvertisingParametersShouldBeValid) {
     advParams.filter_policy = BLE_ADV_FP_ANY;
     advParams.interval      = 100;
     advParams.timeout       = 500; // Advertising for 5 seconds
-    ret = ble_gap_set_advertising_parameters(&advParams, NULL);
+    ret = ble_gap_set_advertising_parameters(&advParams, nullptr);
     assertEqual(ret, 0);
 
-    ret = ble_gap_start_advertising(NULL);
+    ret = ble_gap_start_advertising(nullptr);
     assertEqual(ret, 0);
     ret = ble_gap_is_advertising();
     assertEqual(ret, true);
@@ -316,10 +401,10 @@ test(08_BleChangeAdvertisingParametersDuringAdvertising) {
     advParams.filter_policy = BLE_ADV_FP_ANY;
     advParams.interval      = 100;
     advParams.timeout       = 0; // Advertising forever
-    ret = ble_gap_set_advertising_parameters(&advParams, NULL);
+    ret = ble_gap_set_advertising_parameters(&advParams, nullptr);
     assertEqual(ret, 0);
 
-    ret = ble_gap_start_advertising(NULL);
+    ret = ble_gap_start_advertising(nullptr);
     assertEqual(ret, 0);
     ret = ble_gap_is_advertising();
     assertEqual(ret, true);
@@ -330,7 +415,7 @@ test(08_BleChangeAdvertisingParametersDuringAdvertising) {
     advParams.filter_policy = BLE_ADV_FP_ANY;
     advParams.interval      = 100;
     advParams.timeout       = 500; // Advertising for 5 seconds
-    ret = ble_gap_set_advertising_parameters(&advParams, NULL);
+    ret = ble_gap_set_advertising_parameters(&advParams, nullptr);
     assertEqual(ret, 0);
 
     delay(6000);
@@ -339,7 +424,7 @@ test(08_BleChangeAdvertisingParametersDuringAdvertising) {
     assertEqual(ret, false);
 }
 
-test(09_BleChangeAdvertisingDataDuringAdvertising) {
+test(09_BleChangeAdvertisingDataDuringAdvertising_Next_ConnectUsingApp) {
     int ret;
     uint8_t advDataSet1[] = {
         0x02,
@@ -358,11 +443,11 @@ test(09_BleChangeAdvertisingDataDuringAdvertising) {
         'B','L','E',' ','N','A','M','E','2'
     };
 
-    ret = ble_gap_set_advertising_data(advDataSet1, sizeof(advDataSet1), NULL);
+    ret = ble_gap_set_advertising_data(advDataSet1, sizeof(advDataSet1), nullptr);
     assertEqual(ret, 0);
 
     // Advertising for 5 seconds
-    ret = ble_gap_start_advertising(NULL);
+    ret = ble_gap_start_advertising(nullptr);
     assertEqual(ret, 0);
     ret = ble_gap_is_advertising();
     assertEqual(ret, true);
@@ -370,7 +455,7 @@ test(09_BleChangeAdvertisingDataDuringAdvertising) {
     delay(2500);
 
     // Update the advertising data. It will restart advertising for 5 seconds
-    ret = ble_gap_set_advertising_data(advDataSet2, sizeof(advDataSet1), NULL);
+    ret = ble_gap_set_advertising_data(advDataSet2, sizeof(advDataSet1), nullptr);
     assertEqual(ret, 0);
     ret = ble_gap_is_advertising();
     assertEqual(ret, true);
@@ -383,7 +468,7 @@ test(09_BleChangeAdvertisingDataDuringAdvertising) {
     assertEqual(ret, false);
 }
 
-test(10_BleAddServicesAndCharacteristics_NeedToConnectAndCheckOnTheCentralSide) {
+test(10_BleAddServicesAndCharacteristics_NeedToConnectAndCheckOnTheCentralSide_Next_DisconnectAndReconnectUsingApp) {
     int ret;
     hal_ble_adv_params_t advParams;
 
@@ -393,28 +478,28 @@ test(10_BleAddServicesAndCharacteristics_NeedToConnectAndCheckOnTheCentralSide) 
 
     bleUuid.type = BLE_UUID_TYPE_128BIT;
     memcpy(bleUuid.uuid128, svc1UUID, 16);
-    ble_gatt_server_add_service(BLE_SERVICE_TYPE_PRIMARY, &bleUuid, &svcHandle, NULL);
+    ble_gatt_server_add_service(BLE_SERVICE_TYPE_PRIMARY, &bleUuid, &svcHandle, nullptr);
 
     memset(&char_init, 0x00, sizeof(hal_ble_char_init_t));
     memcpy(bleUuid.uuid128, char1UUID, 16);
     char_init.properties = BLE_SIG_CHAR_PROP_READ|BLE_SIG_CHAR_PROP_WRITE;
     char_init.service_handle = svcHandle;
     char_init.uuid = bleUuid;
-    ble_gatt_server_add_characteristic(&char_init, &characteristic1, NULL);
+    ble_gatt_server_add_characteristic(&char_init, &localCharacteristic1, nullptr);
 
     memcpy(bleUuid.uuid128, svc2UUID, 16);
-    ble_gatt_server_add_service(BLE_SERVICE_TYPE_PRIMARY, &bleUuid, &svcHandle, NULL);
+    ble_gatt_server_add_service(BLE_SERVICE_TYPE_PRIMARY, &bleUuid, &svcHandle, nullptr);
 
     memset(&char_init, 0x00, sizeof(hal_ble_char_init_t));
     memcpy(bleUuid.uuid128, char2UUID, 16);
     char_init.properties = BLE_SIG_CHAR_PROP_NOTIFY;
     char_init.service_handle = svcHandle;
     char_init.uuid = bleUuid;
-    ble_gatt_server_add_characteristic(&char_init, &characteristic2, NULL);
+    ble_gatt_server_add_characteristic(&char_init, &localCharacteristic2, nullptr);
 
     bleUuid.type = BLE_UUID_TYPE_16BIT;
     bleUuid.uuid16 = 0x1234;
-    ble_gatt_server_add_service(BLE_SERVICE_TYPE_PRIMARY, &bleUuid, &svcHandle, NULL);
+    ble_gatt_server_add_service(BLE_SERVICE_TYPE_PRIMARY, &bleUuid, &svcHandle, nullptr);
 
     memset(&char_init, 0x00, sizeof(hal_ble_char_init_t));
     bleUuid.uuid16 = 0x5678;
@@ -422,25 +507,25 @@ test(10_BleAddServicesAndCharacteristics_NeedToConnectAndCheckOnTheCentralSide) 
     char_init.service_handle = svcHandle;
     char_init.uuid = bleUuid;
     char_init.description = "hello";
-    ble_gatt_server_add_characteristic(&char_init, &characteristic3, NULL);
+    ble_gatt_server_add_characteristic(&char_init, &localCharacteristic3, nullptr);
 
     uint8_t data[20] = {0x11};
-    ret = ble_gatt_server_set_characteristic_value(characteristic1.value_handle, data, 5, NULL);
-    assertEqual(ret, 0);
+    ret = ble_gatt_server_set_characteristic_value(localCharacteristic1.value_handle, data, 5, nullptr);
+    assertEqual(ret, 5);
 
     advParams.type          = BLE_ADV_CONNECTABLE_SCANNABLE_UNDIRECRED_EVT;
     advParams.filter_policy = BLE_ADV_FP_ANY;
     advParams.interval      = 100;
     advParams.timeout       = 0; // Advertising forever
-    ret = ble_gap_set_advertising_parameters(&advParams, NULL);
+    ret = ble_gap_set_advertising_parameters(&advParams, nullptr);
     assertEqual(ret, 0);
 
-    ret = ble_gap_start_advertising(NULL);
+    ret = ble_gap_start_advertising(nullptr);
     assertEqual(ret, 0);
 
     // Use a BLE central device to connect to it and check the BLE services and characteristics
     uint16_t timeout = 0;
-    while(!ble_gap_is_connected()) {
+    while(!ble_gap_is_connected(nullptr)) {
         delay(500);
         timeout += 500;
         if(timeout >= 15000) {
@@ -448,7 +533,7 @@ test(10_BleAddServicesAndCharacteristics_NeedToConnectAndCheckOnTheCentralSide) 
         }
     }
 
-    ret = ble_gap_is_connected();
+    ret = ble_gap_is_connected(nullptr);
     assertEqual(ret, true);
 }
 
@@ -458,11 +543,11 @@ test(11_BlePeripheralCanDisconnectInitially_NeedToConnectByCentralSide) {
     // Wait until the previous test case finished.
     // This disconnection should be initiated by the central side.
     // Once disconnected, it should restart advertising automatically.
-    while(ble_gap_is_connected());
+    while(ble_gap_is_connected(nullptr));
 
     // Use a BLE central device to connect to it and wait for 2 seconds without doing anything.
     uint16_t timeout = 0;
-    while(!ble_gap_is_connected()) {
+    while(!ble_gap_is_connected(nullptr)) {
         delay(500);
         timeout += 500;
         if(timeout >= 15000) {
@@ -470,17 +555,155 @@ test(11_BlePeripheralCanDisconnectInitially_NeedToConnectByCentralSide) {
         }
     }
 
-    ret = ble_gap_is_connected();
+    ret = ble_gap_is_connected(nullptr);
     assertEqual(ret, true);
 
     delay(5000);
-    ret = ble_gap_disconnect(0, NULL);
+    ret = ble_gap_disconnect(connHandle, nullptr);
     assertEqual(ret, 0);
 
     // It should restart advertising
-    ret = ble_gap_is_connected();
+    ret = ble_gap_is_connected(nullptr);
     assertEqual(ret, false);
     ret = ble_gap_is_advertising();
     assertEqual(ret, true);
+}
+
+test(12_BleSetScanParametersShouldBeValid) {
+    int ret;
+    hal_ble_scan_params_t scanParams;
+
+    scanParams.active = true;
+    scanParams.filter_policy = BLE_SCAN_FP_ACCEPT_ALL;
+    scanParams.interval = 1;
+    scanParams.window   = 1;
+    scanParams.timeout  = 0; // 0 for forever unless stop initially
+    ret = ble_gap_set_scan_parameters(&scanParams, NULL);
+    assertNotEqual(ret, 0);
+
+    scanParams.interval = 1600; // 1 seconds
+    ret = ble_gap_set_scan_parameters(&scanParams, NULL);
+    assertNotEqual(ret, 0);
+
+    scanParams.window = 1601;
+    ret = ble_gap_set_scan_parameters(&scanParams, NULL);
+    assertNotEqual(ret, 0);
+
+    scanParams.window = 100;
+    ret = ble_gap_set_scan_parameters(&scanParams, NULL);
+    assertEqual(ret, 0);
+}
+
+test(13_BleScanTimeoutAsExpected) {
+    int ret;
+    hal_ble_scan_params_t scanParams;
+
+    scanParams.active = true;
+    scanParams.filter_policy = BLE_SCAN_FP_ACCEPT_ALL;
+    scanParams.interval = 1600;
+    scanParams.window   = 100;
+    scanParams.timeout  = 200; // Scan for 2 seconds
+    ret = ble_gap_set_scan_parameters(&scanParams, NULL);
+    assertEqual(ret, 0);
+
+    // It blocks until the scanning timeout.
+    ret = ble_gap_start_scan(nullptr, nullptr, nullptr);
+    assertEqual(ret, 0);
+
+    ret = ble_gap_is_scanning();
+    assertEqual(ret, false);
+}
+
+test(14_BleConnectToPeerSuccessfully) {
+    int ret;
+    hal_ble_scan_params_t scanParams;
+
+    scanParams.active = true;
+    scanParams.filter_policy = BLE_SCAN_FP_ACCEPT_ALL;
+    scanParams.interval = 1600;
+    scanParams.window   = 100;
+    scanParams.timeout  = 2000; // Scan for 10 seconds
+    ret = ble_gap_set_scan_parameters(&scanParams, NULL);
+    assertEqual(ret, 0);
+
+    ret = ble_gap_start_scan(ble_on_scan_result, nullptr, nullptr);
+    assertEqual(ret, 0);
+    assertEqual(auxPeripheralFound, true);
+
+    ret = ble_gap_connect(&auxPeripheralAddr, nullptr);
+    assertEqual(ret, 0);
+
+    delay(2000);
+    ret = ble_gap_is_connected(nullptr);
+    assertEqual(ret, true);
+}
+
+test(15_BleDiscoverServicesSuccessfully) {
+    int ret;
+
+    ret = ble_gatt_client_discover_all_services(connHandle, ble_on_services_discovered, nullptr, NULL);
+    assertEqual(ret, 0);
+
+    assertEqual(svc1Discovered, true);
+    assertEqual(svc2Discovered, true);
+    assertEqual(svc3Discovered, true);
+}
+
+test(16_BleDiscoverCharacteristicsSuccessfully) {
+    int ret;
+
+    ret = ble_gatt_client_discover_characteristics(connHandle, &service1, ble_on_characteristics_discovered, nullptr, NULL);
+    assertEqual(ret, 0);
+    assertEqual(char1Discovered, true);
+
+    ret = ble_gatt_client_discover_characteristics(connHandle, &service2, ble_on_characteristics_discovered, nullptr, NULL);
+    assertEqual(ret, 0);
+    assertEqual(char2Discovered, true);
+
+    ret = ble_gatt_client_discover_characteristics(connHandle, &service3, ble_on_characteristics_discovered, nullptr, NULL);
+    assertEqual(ret, 0);
+    assertEqual(char3Discovered, true);
+}
+
+test(17_BleReadWriteDataSuccessfully) {
+    int ret;
+    uint8_t readData[20];
+
+    ret = ble_gatt_client_read(connHandle, char1.value_handle, readData, sizeof(readData), NULL);
+    assertEqual(ret, 5); // Initial value length of this characteristic of Auxiliary peripheral.
+
+    uint8_t writeData[] = {0xaa, 0xbb, 0x5a, 0xa5};
+    ret = ble_gatt_client_write_with_response(connHandle, char1.value_handle, writeData, sizeof(writeData), NULL);
+    assertEqual(ret, 4);
+
+    ret = ble_gatt_client_read(connHandle, char1.value_handle, readData, sizeof(readData), NULL);
+    assertEqual(ret, 4);
+    ret = memcmp(readData, writeData, sizeof(writeData));
+    assertEqual(ret, 0);
+
+    ret = ble_gatt_client_write_without_response(connHandle, char3.value_handle, writeData, sizeof(writeData), NULL);
+    assertEqual(ret, 4);
+}
+
+test(18_BleConfigCccdSuccessfully) {
+    int ret;
+
+    ret = ble_gatt_client_configure_cccd(connHandle, char2.cccd_handle, BLE_SIG_CCCD_VAL_NOTIFICATION, NULL);
+    assertEqual(ret, 0);
+
+    delay(5000);
+    assertEqual(notifiedData, true);
+
+    ret = ble_gatt_client_configure_cccd(connHandle, char2.cccd_handle, BLE_SIG_CCCD_VAL_DISABLED, NULL);
+    assertEqual(ret, 0);
+}
+
+test(19_BleDisconnectSuccessfully) {
+    int ret;
+
+    ret = ble_gap_disconnect(connHandle, NULL);
+    assertEqual(ret, 0);
+    ret = ble_gap_is_connected(nullptr);
+    assertEqual(ret, false);
 }
